// packages/www/src/types/playoffs.ts
export interface NHLDivision {
  id: number
  name: string
  nameShort: string
  link: string
  abbreviation: string
  conference: {
    id: number
    name: string
    link: string
  }
  active: boolean
}